Add book cover uploading functionality
This commit introduces the option to upload and save book covers. A new endpoint "/add/cover" has been added to handle book cover uploads. In the event a book does not have a cover, an upload input field is provided in the books display page. An environment variable is now utilized to set the pathway of the cover folder.