Add app name and version to page title
The application name and version have been added to the page title in the base.html template. This used new Jinja global variables "app_name" and "version" which were set in the application's Jinja environment within the core's app.py module. The version is also displayed in the navigation bar.